pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "serpapi-mcp-server"
version = "0.2.0"
description = "A Model Context Protocol (MCP) server implementation that integrates with SerpApi for comprehensive search engine results and data extraction"
readme = "README.md"
license = "MIT"
authors = [
    { name = "SerpApi", email = "hello@serpapi.com" }
]
keywords = ["mcp", "serpapi", "search", "api", "model-context-protocol"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: Indexing/Search",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.3.0",
    "google-search-results>=2.4.2",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
]

[project.urls]
Homepage = "https://github.com/ilyazub/serpapi-mcp-server"
Repository = "https://github.com/ilyazub/serpapi-mcp-server"
Issues = "https://github.com/ilyazub/serpapi-mcp-server/issues"
Documentation = "https://github.com/ilyazub/serpapi-mcp-server#readme"

[project.scripts]
serpapi-mcp = "src.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py313']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
